requirements.txt
paho-mqtt==1.5.0
ruuvitag-sensor==0.3.3
pyyaml==5.4.1
influxdb==5.3.0
